parallel_openai/requirements.txt
openai>=1.0.0
pyyaml>=6.0
aiohttp>=3.8.0